+++ src/renderer/runtime.ts
@@ -1,10 +1,8 @@
 import { ipcBackend } from './ipc'
 import { RC_Config } from '../shared/shared-types'
-import { getLogger, setLogHandler } from '../shared/logger'
+import { setLogHandler } from '../shared/logger'
 import type { dialog, app, shell, clipboard } from 'electron'
 import { basename, join } from 'path'
-
-const log = getLogger('renderer/runtime')
 
 const {
   openExternal,
@@ -128,11 +126,7 @@ class Electron implements Runtime {
   }
   private rc_config: RC_Config = null
   transformBlobURL(blob: string): string {
-    if (!blob) {
-      log.warn('transformBlobURL: received a null/undefined blob string ')
-      return null
-    }
-    return 'dc-blob://' + blob.substring(blob.indexOf('accounts') + 9)
+    return blob
   }
   async showOpenFileDialog(
     options: Electron.OpenDialogOptions