+++ src/main/unread-badge.ts
@@ -16,7 +16,7 @@ export default function setupUnreadBadge(dc: DeltaChatController) {
   }
 
   dc._dc.on('DC_EVENT_INCOMING_MSG', (_chatId: number, _msgId: number) => {
-    // don't update imidiately if the app is in focused
+    // don't update immediately if the app is in focused
     if (mainWindow.window.hidden) update()
 
     // update after a delay again to make sure its up to date