+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 ### Fixed
 - Fix dragging out file attachments #2177
+- own location and path in maps is not visible in single chats
+
 
 ### Changed
 - Switch back to file scheme #2171 fixes audio and video seeking issues
+- increase `DAYS_UNTIL_UPDATE_SUGGESTION` to 120 days
 
 ## [1.15.3] - 2021-03-18