+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [Unreleased][unreleased]
 
 ### Fixed
+- Fix dragging out file attachments #2177
 - own location and path in maps is not visible in single chats
+
 
 ### Changed
 - Switch back to file scheme #2171 fixes audio and video seeking issues