+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## [Unreleased][unreleased]
 
+### Fixed
+- Fix dragging out file attachments #2177
+- own location and path in maps is not visible in single chats
+
+
 ### Changed
 - Switch back to file scheme #2171 fixes audio and video seeking issues
 - adjust warning/hint color of before-login-hint
+- increase `DAYS_UNTIL_UPDATE_SUGGESTION` to 120 days
 
 ## [1.15.3] - 2021-03-18