+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 - Switch back to file scheme #2171 fixes audio and video seeking issues
+- adjust warning/hint color of before-login-hint
 - increase `DAYS_UNTIL_UPDATE_SUGGESTION` to 120 days
 
 ## [1.15.3] - 2021-03-18