+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix dragging out file attachments #2177
 - own location and path in maps is not visible in single chats
 - Fix crash in settings-profile if account object is empty
+- make notifications more reliable and allow multiple ones at a time
 
 ### Changed
 - Switch back to file scheme #2171 fixes audio and video seeking issues