+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Switch back to file scheme #2171 fixes audio and video seeking issues
 - adjust warning/hint color of before-login-hint
 - increase `DAYS_UNTIL_UPDATE_SUGGESTION` to 120 days
+- improve notifications (show chat name and avatar)
 
 ## [1.15.3] - 2021-03-18